(MAINT) include role::by_size::small instead of duplicating code
Prior to this commit, we included the same code in the medium size
as the small size.

After this commit, we simply include the small size and avoid the
duplication.  Each larger size should build on top of smaller sizes
and changes to smaller sizes shouldn't need to be copied around.

This breaks the convention of not including roles in other roles
but I think it's fine for this use case.